Add SignUpSchema file  with validations of register form
